tests/wpt/tests/speculation-rules/prerender/cross-origin-isolated.https.html

<!DOCTYPE html>
<title>Allow crossOriginIsolated in prerendered page</title>
<meta name="variant" content="?target_hint=_self">
<meta name="variant" content="?target_hint=_blank">
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="../resources/utils.js"></script>
<script src="resources/utils.js"></script>

<body>
<script>
setup(() => assertSpeculationRulesIsSupported());

promise_test(async t => {
  const uid = token();
  const testChannel = new PrerenderChannel('test-channel', uid);
  t.add_cleanup(() => {
    testChannel.close();
  });
  const gotMessage = new Promise(resolve => {
    testChannel.addEventListener('message', e => resolve(e.data), {once: true});
  });

  startPrerendering(`resources/cross-origin-isolated.https.html?uid=${uid}&target_hint=${getTargetHint()}`);
  assert_true(await gotMessage);
}, 'Allow crossOriginIsolated in prerendered page');
</script>
